achcli: introduce 'describe' package for sharing human-readable formatting
