+++ cmd/achcli/describe/code_map.go
@@ -1,4 +1,8 @@
-package main
+// Copyright 2020 The Moov Authors
+// Use of this source code is governed by an Apache License
+// license that can be found in the LICENSE file.
+
+package describe
 
 import (
 	"fmt"
